scripts/requirements.txt
numpy
python-dotenv
biopython
pyarrow
pandas
termcolor
argparse
pathlib
fastparquet
